Initial commit of lesson 6's active port
While this code functions to complete the lesson, I don't like the design of the ActiveBuzzer struct. There may be a technique to precisely dial in the desired output sound based on delay timing, but I don't know what that is and frankly, my ears can't handle any further testing.